+++ miden-crypto/src/merkle/smt/full/mod.rs
@@ -96,7 +96,11 @@ impl Smt {
         }
     }
 
-    /// Similar to `with_entries` but used for pre-sorted entries to avoid overhead
+    /// Similar to `with_entries` but avoids the overhead of sorting if the entries are already
+    /// sorted.
+    ///
+    /// This only applies if the "concurrent" feature is enabled. Without the feature, the behavior
+    /// is equivalent to `with_entiries`.
     pub fn with_sorted_entries(
         entries: impl IntoIterator<Item = (RpoDigest, Word)>,
     ) -> Result<Self, MerkleError> {
@@ -578,7 +582,7 @@ fn test_smt_serialization_deserialization() {
 }
 
 #[test]
-fn test_smt_sorted_serialization_deserialization() {
+fn smt_with_sorted_entries() {
     // Smt with sorted values
     let smt_leaves_2: [(RpoDigest, Word); 2] = [
         (
@@ -590,29 +594,9 @@ fn test_smt_sorted_serialization_deserialization() {
             [Felt::new(5_u64), Felt::new(6_u64), Felt::new(7_u64), Felt::new(8_u64)],
         ),
     ];
+
     let smt = Smt::with_sorted_entries(smt_leaves_2).unwrap();
+    let expected_smt = Smt::with_entries(smt_leaves_2).unwrap();
 
-    let bytes = smt.to_bytes();
-    assert_eq!(smt, Smt::read_from_bytes(&bytes).unwrap());
-    assert_eq!(bytes.len(), smt.get_size_hint());
-}
-
-#[test]
-fn test_smt_sorted_serialization_deserialization_fail() {
-    // Smt with values
-    let smt_leaves_2: [(RpoDigest, Word); 2] = [
-        (
-            RpoDigest::new([Felt::new(105), Felt::new(106), Felt::new(107), Felt::new(108)]),
-            [Felt::new(5_u64), Felt::new(6_u64), Felt::new(7_u64), Felt::new(8_u64)],
-        ),
-        (
-            RpoDigest::new([Felt::new(101), Felt::new(102), Felt::new(103), Felt::new(104)]),
-            [Felt::new(1_u64), Felt::new(2_u64), Felt::new(3_u64), Felt::new(4_u64)],
-        ),
-    ];
-    let result = std::panic::catch_unwind(|| {
-        let _ = Smt::with_sorted_entries(smt_leaves_2);
-    });
-
-    assert!(result.is_err(), "Expected panic for unsorted entries");
+    assert_eq!(smt, expected_smt);
 }